Upload All Changes
Better security Check for gettings Files form the API
Better Streaming Mode
Caching now works on ('/tts_stream')
Add ("/get_stream_mode/") which returns, if the stream_mode is enabled...
Add check if speaker was not found
Make the voice generation for ('/tts_stream') synchronized, so the model doesn't crash on multi threading calls.
Add System Variable to pick the audiodevice, XTTS uses for streaming_mode

+++ xtts_api_server/tts_funcs.py
@@ -91,9 +91,11 @@ class TTSWrapper:
         self.is_official_model = True
         
         if self.enable_cache_results:
-            # Reset the contents of the cache file at each initialization.
-            with open(self.cache_file_path, 'w') as cache_file:
-                json.dump({}, cache_file)
+            if not os.path.isfile(self.cache_file_path):
+                #only recreate, when file is not found
+                with open(self.cache_file_path, 'w') as cache_file:
+                    json.dump({}, cache_file)
+                
     # HELP FUNC
     def isModelOfficial(self,model_version):
         if model_version in official_model_list:
